requirements.txt
pyserial>=3.5
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
google-generativeai>=0.3.0
gtts>=2.4.0
playsound>=1.3.0
python-dotenv>=1.0.0
pyttsx3>=2.90
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0

